update: utils function created to retrieve stats
